docker_images/deepst/requirements.txt
# DeepST Requirements
# Core scientific computing
numpy>=1.21.0,<2.0.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Image processing (for morphological features)
torchvision>=0.14.0
Pillow>=9.0.0

# Single-cell and spatial transcriptomics
scanpy>=1.10.0
igraph>=0.11.0
leidenalg>=0.10.0
louvain>=0.8.0

# Configuration and utilities
pyyaml>=5.4.0
tqdm>=4.62.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# LangChain integration (for AutoDRP framework)
langchain>=0.0.1
langchain-core>=0.0.1
python-dotenv>=0.19.0
langchain-mcp-adapters>=0.0.1
prettytable>=3.0.0
